requirements.txt

Flask
pillow
pytesseract
